projects/image-api/pyproject.toml
[project]
name = "image-api"
version = "0.1.0"
description = "Self-hosted image optimization API"
readme = "README.md"
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.26.0",
    "pillow>=10.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["app/__init__.py"]